refactor(api): restructure backend
- use lib.rs
- move database models into own crate
