+++ crates/lfu_database/src/lib.rs
@@ -1,0 +1,4 @@
+#![doc = include_str!("../README.md")]
+
+pub mod models;
+pub mod schema;
