+++ crates/lfu_database/src/models/mod.rs
