+++ crates/lfu_database/src/schema.rs
