+++ diesel.toml
@@ -1,9 +1,5 @@
 # For documentation on how to configure this file,
-# see https://diesel.rs/guides/configuring-diesel-cli
-
+# see diesel.rs/guides/configuring-diesel-cli
 [print_schema]
-file = "src/schema.rs"
+file = "crates/lfu_database/src/schema.rs"
 custom_type_derives = ["diesel::query_builder::QueryId"]
-
-[migrations_directory]
-dir = "migrations"
