+++ docs/API.md
@@ -2,7 +2,7 @@
 
 ## Server
 
-The code to run the server is located in _src/main.rs_. This is where the system is pieced together and instantiated, and can be seen as the "entry point" to learn.framer.university.
+The code to run the server is located in _src/main.rs_. This is where the system is pieced together and instantiated, and can be seen as the "entry point" to api.learn.framer.university.
 
 The server does the following:
 
@@ -58,22 +58,6 @@ This module is responsible for managing database connections and migrations.
 ### `model`
 
 This module contains the data models used in the application. These models represent the structure of the data stored in the database and are defined using Diesel ORM.
-
-### `schema.rs`
-
-This module contains the schema definitions for the database tables. These definitions are generated by Diesel and are used to map Rust structs to database tables.
-
-- **Schema Definitions**: Automatically generated by Diesel based on the database schema.
-- **Table Definitions**: Includes table definitions for users, courses, enrollments, etc.
-
-## Database
-
-After creating new migrations (via `diesel migration generate`), you can update
-the schema by running:
-
-```sh
-diesel print-schema > src/schema.rs
-```
 
 ## Tests
