+++ src/config/mod.rs
@@ -2,5 +2,4 @@ mod sentry;
 mod server;
 
 pub use self::sentry::SentryConfig;
-pub use self::server::Env;
 pub use self::server::Server;
