+++ src/tests/util/mod.rs
@@ -1,8 +1,11 @@
-pub mod test_app;
-
-use crate::{auth::Tokens, models::user::User};
 use axum_test::{TestRequest, TestServer};
-pub use test_app::TestApp;
+
+use lfu_database::models::user::User;
+
+use crate::auth::Tokens;
+
+pub mod test_app;
+use test_app::TestApp;
 
 /// A collection of helper methods for the three authentication types
 /// - Anonymous
